CONTRAST-8163 complete refactor of models and urls
