+++ src/main/java/com/contrastsecurity/models/AgentType.java
@@ -1,4 +1,4 @@
-package com.contrastsecurity.rest;
+package com.contrastsecurity.models;
 
 /**
  * Enumerate the agent downloads.
