+++ src/main/java/com/contrastsecurity/models/CodeMethod.java
@@ -26,13 +26,28 @@
  * STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF
  * THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package com.contrastsecurity.rest;
+package com.contrastsecurity.models;
 
-public class UnauthorizedException extends Exception {
-	
-	public UnauthorizedException(int rc) {
-		super("Received response code: " + rc);
+/**
+ * An application method.
+ */
+public class CodeMethod {
+
+	/**
+	 * Return the signature of this method. 
+	 * @return the method signature
+	 */
+	public String getSignature() {
+		return signature; 
 	}
+	private String signature;
 
-	private static final long serialVersionUID = -9049287248312255189L;
+	/**
+	 * Return whether or not this method has been executed.
+	 * @return true, if this method has been executed; false otherwise
+	 */
+	public boolean isExecuted() { 
+		return executed;
+	}
+	private boolean executed;
 }
