+++ src/main/java/com/contrastsecurity/models/PlatformVulnerability.java
@@ -26,55 +26,46 @@
  * STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF
  * THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package com.contrastsecurity.rest;
+package com.contrastsecurity.models;
 
-import java.util.List;
-
-import com.google.gson.annotations.SerializedName;
-
-/**
- * An HTTP request.
- */
-public class HttpRequest {
-
-	public int getPort() {
-		return port;
-	}
-	private int port;
+public class PlatformVulnerability {
 	
-	public String getProtocol() {
-		return protocol;
+	/**
+	 * Return the CVE associated with this vulnerability.
+	 * 
+	 * @return the CVE associated with this vulnerability
+	 */
+	public String getCve() {
+		return cve;
 	}
-	private String protocol;
+	private String cve;
 	
-	public String getMethod() {
-		return method;
+	/**
+	 * Return the severity of the CVE; one of Critical, High, Medium, Low, Note.
+	 * 
+	 * @return the severity of the CVE; one of Critical, High, Medium, Low, Note 
+	 */
+	public String getSeverity() {
+		return severity;
 	}
-	private String method;
+	private String severity;
 	
-	public String getUri() {
-		return uri;
+	/**
+	 * Return a short description about the issue.
+	 * 
+	 * @return a short description about the issue
+	 */
+	public String getDescription() {
+		return description;
 	}
-	private String uri;
+	private String description;
 	
-	public String getVersion() {
-		return version;
+	/**
+	 * Return a link to the issue from the vendor.
+	 * @return a link to the issue from the vendor
+	 */
+	public String getLink() {
+		return link;
 	}
-	private String version;
-	
-	public String getQueryString() {
-		return queryString;
-	}
-	@SerializedName("query-string")
-	private String queryString;
-	
-	public List<NameValuePair> getHeaders() {
-		return headers;
-	}
-	private List<NameValuePair> headers;
-	
-	public List<NameValuePair> getParameters() {
-		return parameters;
-	}
-	private List<NameValuePair> parameters;
+	private String link;
 }
