+++ src/main/java/com/contrastsecurity/models/Trace.java
@@ -26,8 +26,10 @@
  * STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF
  * THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package com.contrastsecurity.rest;
+package com.contrastsecurity.models;
 
+import com.contrastsecurity.models.HttpRequest;
+import com.contrastsecurity.rest.TraceEvent;
 import com.google.gson.annotations.SerializedName;
 
 import java.util.List;
