+++ src/main/java/com/contrastsecurity/models/TraceEvent.java
@@ -26,7 +26,9 @@
  * STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF
  * THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package com.contrastsecurity.rest;
+package com.contrastsecurity.models;
+
+import com.contrastsecurity.models.CodeObject;
 
 import java.util.List;
