+++ src/main/java/com/contrastsecurity/sdk/http/UrlBuilder.java
@@ -1,0 +1,7 @@
+package com.contrastsecurity.sdk.http;
+
+/**
+ * Created by justinleo on 4/21/16.
+ */
+public class UrlBuilder {
+}
